type changes

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: mwojtasi <mwojtasi@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 15:18:26 by mwojtasi          #+#    #+#             */
-/*   Updated: 2023/11/08 05:44:40 by mwojtasi         ###   ########.fr       */
+/*   Updated: 2023/11/23 01:02:01 by mwojtasi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != (unsigned char)c && s[i])
